+++ astropy/timeseries/tests/test_sampled.py
@@ -10,7 +10,9 @@ from astropy.table import Table, Column
 from astropy.time import Time, TimeDelta
 from astropy import units as u
 from astropy.utils.data import get_pkg_data_filename
+from astropy.tests.helper import assert_quantity_allclose
 
+from astropy.timeseries.periodograms import BoxLeastSquares, LombScargle
 from astropy.timeseries.sampled import TimeSeries
 
 INPUT_TIME = Time(['2016-03-22T12:30:31',
@@ -277,3 +279,27 @@ def test_required_columns():
         ts.copy().rename_column('time', 'banana')
     assert exc.value.args[0] == ("TimeSeries object is invalid - expected "
                                  "'time' as the first column but found 'banana'")
+
+
+@pytest.mark.parametrize('cls', [BoxLeastSquares, LombScargle])
+def test_periodogram(cls):
+
+    # Note that we don't need to check the actual results from the periodogram
+    # classes here since these are tested extensively in
+    # astropy.timeseries.periodograms.
+
+    ts = TimeSeries(time=INPUT_TIME,
+                    data=[[10, 2, 3], [4, 5, 6]],
+                    names=['a', 'b'])
+
+    p1 = cls.from_timeseries(ts, 'a')
+    assert isinstance(p1, cls)
+    assert_allclose(p1.t.jd, ts.time.jd)
+    assert_equal(p1.y, ts['a'])
+    assert p1.dy is None
+
+    p2 = cls.from_timeseries(ts, 'a', uncertainty='b')
+    assert_quantity_allclose(p2.dy, ts['b'])
+
+    p3 = cls.from_timeseries(ts, 'a', uncertainty=0.1)
+    assert_allclose(p3.dy, 0.1)